+++ AngularClient/src/main/core/components/landing-page.component.ts
@@ -6,18 +6,10 @@ import { User, AuthService } from "backbone-client-core";
   templateUrl: "landing-page.component.html",
 })
 export class LandingPageComponent {
+
   currentUser: User = null;
 
   constructor(private authService: AuthService) {
     this.currentUser = this.authService.currentUser;
   }
-
-  isLogedIn(): boolean {
-    if (this.authService.currentUser === null) {
-      return false;
-    } else {
-      return this.authService.currentUser.isAuthenticated();
-    }
-  }
-
 }