+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### Changelog
 
+**0.9.15**
+
+* AngularClient - LoginReturnUrl fix
+
 **0.9.14**
 
 * AngularClient - Project manager minor fix